Get mem tracker to working state. Debug menu next. Other small random fixes.
