+++ DebugUI/DebugUI.cpp
@@ -1,15 +1,10 @@
 #include "DebugUI.h"
 #include "CoreDefines.h"
-#include "DataStructures/HashMap.h"
 #include "DataStructures/Vector.h"
 #include "Game/GraphicsTypes.h"
-#include "Graphics/Common/GPUTimestamps.h"
 #include "Graphics/Common/GraphicsCommon.h"
-#include "Hashing.h"
 #include "Math/VectorTypes.h"
 #include "Platform/PlatformGameAPI.h"
-#include "Sorting.h"
-#include "StringTypes.h"
 #include "ThirdParty/imgui-docking/imgui.h"
 
 static const uint32 MAX_VERTS = 1024 * 1024;
@@ -36,13 +31,6 @@ namespace DebugUI
   static Tk::Graphics::MemoryMappedBufferPtr colorBufPtr = {};
   static uint32 accumVtxOffset = 0;
   static uint32 accumIdxOffset = 0;
-
-  static bool g_enable = false;
-
-  void ToggleEnable()
-  {
-    g_enable = !g_enable;
-  }
 
   void* ImGuiMemWrapper_Malloc(size_t sz, void* user_data)
   {
@@ -477,319 +465,6 @@ namespace DebugUI
           graphicsCommandStream->Clear();
         }
       }
-    }
-  }
-
-  static bool mainMenu_SelectedOverview = false;
-  static bool mainMenu_SelectedRPTimings = false;
-
-  void UI_MainMenu()
-  {
-    using namespace Tk;
-    using namespace Graphics;
-
-    if (!g_enable)
-    {
-      mainMenu_SelectedOverview = false;
-      mainMenu_SelectedRPTimings = false;
-      return;
-    }
-
-    if (ImGui::BeginMainMenuBar())
-    {
-      if (ImGui::BeginMenu("Performance"))
-      {
-        if (ImGui::MenuItem("Overview", NULL, &mainMenu_SelectedOverview))
-        {
-        }
-        if (ImGui::MenuItem("GPU Render Pass Timings", NULL, &mainMenu_SelectedRPTimings))
-        {
-        }
-
-        ImGui::EndMenu();
-      }
-    }
-    ImGui::EndMainMenuBar();
-  }
-
-  void UI_PerformanceOverview()
-  {
-    using namespace Tk;
-    using namespace Graphics;
-
-    if (mainMenu_SelectedOverview)
-    {
-      if (ImGui::Begin("Performance Overview"))
-      {
-        GPUTimestamps::TimestampData timestampData = GPUTimestamps::GetTimestampData();
-        ImGui::Text("%s: %.2f\n", "Total Frame Time", timestampData.totalFrameTimeInUS);
-      }
-      ImGui::End();
-    }
-  }
-
-  void UI_RenderPassStats()
-  {
-    using namespace Tk;
-    using namespace Graphics;
-
-    struct RunningTimestampEntry
-    {
-      uint32 numSamples = 0;
-      float runningTermQ = 0.0f;
-      float runningAvg = 0.0f;
-      float runningMax = 0.0f;
-    };
-
-    struct DisplayTimestampEntry
-    {
-      enum : uint8
-      {
-        TimeCurr,
-        TimeAvg,
-        StdDev,
-        TimeMax,
-        DisplayCount,
-      };
-
-      const char* name = NULL;
-      float timeData[DisplayCount] = {};
-    };
-
-    // Track timestamp name hash to running statistics data
-    static const uint32 ReserveEles = 256;
-    static Tk::Core::HashMap<uint64, RunningTimestampEntry, MapHashFn64> runningStatsMap;
-    runningStatsMap.Reserve(ReserveEles);
-
-    // Final list of entries to display for sorting
-    static Tk::Core::Vector<DisplayTimestampEntry> entryDisplayList;
-    entryDisplayList.Reserve(ReserveEles);
-    entryDisplayList.Clear();
-
-    if (mainMenu_SelectedRPTimings)
-    {
-      if (ImGui::Begin("GPU Render Pass Timings", NULL,
-                       ImGuiWindowFlags_AlwaysAutoResize))
-      {
-        static bool shouldCopyToClip = false;
-        static int displayFactorBtnIdx = 0;
-        static float displayConversionFactor = 1.0f;
-
-        if (ImGui::SmallButton("Clear"))
-        {
-          runningStatsMap.Clear();
-        }
-        ImGui::SameLine();
-        if (ImGui::SmallButton("Copy"))
-        {
-          shouldCopyToClip = true;
-        }
-        ImGui::SameLine();
-        if (ImGui::RadioButton("US", &displayFactorBtnIdx, 0))
-        {
-          displayConversionFactor = 1.0f;
-        }
-        ImGui::SameLine();
-        if (ImGui::RadioButton("MS", &displayFactorBtnIdx, 1))
-        {
-          displayConversionFactor = 0.001f;
-        }
-
-        ImGuiTableFlags_ tableFlags = (ImGuiTableFlags_)(ImGuiTableFlags_RowBg
-                                                         | ImGuiTableFlags_SizingFixedSame
-                                                         | ImGuiTableFlags_PadOuterX
-                                                         | ImGuiTableFlags_Resizable
-                                                         | ImGuiTableFlags_Sortable
-                                                         | ImGuiTableFlags_SortTristate);
-
-        const uint32 numCols = 5;
-        if (ImGui::BeginTable("GPU Render Pass Timings Table", numCols, tableFlags))
-        {
-          ImGui::PushStyleColor(ImGuiCol_TableHeaderBg, ImVec4(0.4f, 0.3f, 0.0f, 0.2f));
-          ImGui::PushStyleColor(ImGuiCol_TableRowBg, ImVec4(0.2f, 0.2f, 0.2f, 0.2f));
-          ImGui::PushStyleColor(ImGuiCol_TableRowBgAlt, ImVec4(0.4f, 0.3f, 0.0f, 0.2f));
-          ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.0f, 1.0f, 1.0f, 0.5f));
-
-          const char* headerStrings[numCols] = {
-            "Pass name", "Curr time", "Avg time", "Std dev", "Max time",
-          };
-
-          // Column headers
-          ImGui::TableSetupColumn(headerStrings[0],
-                                  ImGuiTableColumnFlags_PreferSortAscending);
-          ImGui::TableSetupColumn(headerStrings[1],
-                                  ImGuiTableColumnFlags_PreferSortDescending);
-          ImGui::TableSetupColumn(headerStrings[2],
-                                  ImGuiTableColumnFlags_PreferSortDescending);
-          ImGui::TableSetupColumn(headerStrings[3],
-                                  ImGuiTableColumnFlags_PreferSortDescending);
-          ImGui::TableSetupColumn(headerStrings[4],
-                                  ImGuiTableColumnFlags_PreferSortDescending);
-          ImGui::TableHeadersRow();
-
-          // Timestamp data rows
-          GPUTimestamps::TimestampData timestampData = GPUTimestamps::GetTimestampData();
-          for (uint32 i = 0; i < timestampData.numTimestamps; ++i)
-          {
-            const GPUTimestamps::Timestamp& currTimestamp = timestampData.timestamps[i];
-
-            if (!currTimestamp.name)
-            {
-              continue;
-            }
-
-            const Tk::Core::Hash timestampNameHash =
-              HASH_64_RUNTIME(currTimestamp.name, (uint32)strlen(currTimestamp.name));
-
-            RunningTimestampEntry* entry = NULL;
-            uint32 index = runningStatsMap.FindIndex(timestampNameHash.m_val);
-
-            if (index == Tk::Core::HashMapBase::eInvalidIndex)
-            {
-              // First time add to map
-              index = runningStatsMap.Insert(timestampNameHash.m_val, {});
-            }
-            entry = &(runningStatsMap.DataAtIndex(index));
-
-            // Update stats in entry
-            const float currentSample = currTimestamp.timeInst;
-            const float prevRunningAvg = entry->runningAvg;
-            entry->numSamples++;
-            entry->runningMax = Max(entry->runningMax, currentSample);
-            // https://en.wikipedia.org/wiki/Standard_deviation#Rapid_calculation_methods
-            entry->runningAvg =
-              prevRunningAvg + ((currentSample - prevRunningAvg) / entry->numSamples);
-            entry->runningTermQ =
-              entry->runningTermQ
-              + (currentSample - prevRunningAvg) * (currentSample - entry->runningAvg);
-            float currStdDev = 0.0f;
-            if (entry->numSamples > 1)
-            {
-              currStdDev = sqrtf(entry->runningTermQ / (entry->numSamples - 1));
-            }
-
-            DisplayTimestampEntry displayEntry = {};
-            displayEntry.name = currTimestamp.name;
-            displayEntry.timeData[DisplayTimestampEntry::TimeCurr] = currentSample;
-            displayEntry.timeData[DisplayTimestampEntry::TimeAvg] = entry->runningAvg;
-            displayEntry.timeData[DisplayTimestampEntry::StdDev] = currStdDev;
-            displayEntry.timeData[DisplayTimestampEntry::TimeMax] = entry->runningMax;
-            entryDisplayList.PushBackRaw(displayEntry);
-          }
-
-          ImGuiTableSortSpecs* sortSpecs = ImGui::TableGetSortSpecs();
-          if (sortSpecs->SpecsCount)
-          {
-            const ImGuiTableColumnSortSpecs* tableSortSpecs = sortSpecs->Specs;
-
-            Tk::Core::MergeSort(
-              (DisplayTimestampEntry*)entryDisplayList.Data(), entryDisplayList.Size(),
-              [=](const void* A, const void* B)
-              {
-                const DisplayTimestampEntry* entryA = (DisplayTimestampEntry*)A;
-                const DisplayTimestampEntry* entryB = (DisplayTimestampEntry*)B;
-
-                bool compareResult = 0;
-                switch (tableSortSpecs->ColumnIndex)
-                {
-                  case 0:
-                  {
-                    compareResult = strcmp(entryA->name, entryB->name) < 0 ? 1 : 0;
-                    break;
-                  }
-                  case 1:
-                  case 2:
-                  case 3:
-                  case 4:
-                  {
-                    compareResult = entryA->timeData[tableSortSpecs->ColumnIndex - 1]
-                                    < entryB->timeData[tableSortSpecs->ColumnIndex - 1];
-                    break;
-                  }
-
-                  default:
-                  {
-                    break;
-                  }
-                }
-
-                if (tableSortSpecs->SortDirection == ImGuiSortDirection_Descending)
-                {
-                  compareResult = !compareResult;
-                }
-
-                return compareResult;
-              });
-          }
-
-          if (shouldCopyToClip)
-          {
-            static Tk::Core::StrFixedBuffer<1'048'576> csvOutput;
-            csvOutput.Clear();
-
-            // Headers
-            const char* delimiter = ",";
-            for (uint32 uiValue = 0; uiValue < numCols; ++uiValue)
-            {
-              csvOutput.Append(headerStrings[uiValue]);
-              csvOutput.Append(delimiter);
-            }
-            csvOutput.Append("\n");
-
-            // Data
-            for (uint32 i = 0; i < entryDisplayList.Size(); ++i)
-            {
-              const DisplayTimestampEntry& displayEntry = entryDisplayList[i];
-
-              csvOutput.Append(displayEntry.name);
-              csvOutput.Append(delimiter);
-              for (uint32 uiValue = 0; uiValue < DisplayTimestampEntry::DisplayCount;
-                   ++uiValue)
-              {
-                int result = sprintf_s(csvOutput.EndOfStrPtr(), csvOutput.LenRemaining(),
-                                       "%.2f", displayEntry.timeData[uiValue]);
-                TINKER_ASSERT(result != -1);
-                csvOutput.m_len += result;
-                csvOutput.Append(delimiter);
-              }
-              csvOutput.Append("\n");
-            }
-            csvOutput.NullTerminate();
-            ImGui::SetClipboardText(csvOutput.m_data);
-
-            shouldCopyToClip = false;
-          }
-
-          for (uint32 i = 0; i < entryDisplayList.Size(); ++i)
-          {
-            const DisplayTimestampEntry& displayEntry = entryDisplayList[i];
-
-            ImGui::TableNextRow();
-            ImGui::TableNextColumn();
-            ImGui::Text("%s", displayEntry.name);
-            ImGui::TableNextColumn();
-            ImGui::Text("%.3f", displayEntry.timeData[DisplayTimestampEntry::TimeCurr]
-                                  * displayConversionFactor);
-            ImGui::TableNextColumn();
-            ImGui::Text("%.3f", displayEntry.timeData[DisplayTimestampEntry::TimeAvg]
-                                  * displayConversionFactor);
-            ImGui::TableNextColumn();
-            ImGui::Text((const char*)u8"± %.2f",
-                        displayEntry.timeData[DisplayTimestampEntry::StdDev]
-                          * displayConversionFactor);
-            ImGui::TableNextColumn();
-            ImGui::Text("%.3f", displayEntry.timeData[DisplayTimestampEntry::TimeMax]
-                                  * displayConversionFactor);
-          }
-
-          ImGui::EndTable();
-          ImGui::PopStyleColor();
-          ImGui::PopStyleColor();
-          ImGui::PopStyleColor();
-          ImGui::PopStyleColor();
-        }
-      }
-      ImGui::End();
     }
   }
 } //namespace DebugUI
