+++ Game/GameDebugMenus.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void UpdateAllDebugMenus();
+void ToggleEnable();
